Use steno to have atomic file writes
fs.writeFile is not atomic, in a way, first an empty file is created and
then the file is written to in buffers. steno.writeFile instead creates
the file in a temp location and then renames it to the location that we
want the to be in the cache. This way at any moment the cache file
wouldn't be an empty or invalid json file.

The test was removed due to the fact that file permissions don't matter
in unix when removing files, and the directory permission is taken into
account now. Since now we are using steno, and it renames the file into
the location, as long as it has write permission for the directory it
can rename the file.